requirements.txt
PyQt5>=5.8.2, <=5.14.0
numpy
PythonQwt